moved generic datetime to logger
