+++ src/handlers/Request.ts
@@ -2,6 +2,7 @@ import { CryptoUtils } from '../crypto/Utils';
 import { RequestValidator } from '../templates/RequestValidator';
 import Cache from '../Cache';
 import axios from 'axios';
+import { dateTimeLogger } from '../logger/Logger';
 
 var cryptoUtils = new CryptoUtils();
 var dateFormat = require('dateformat');
@@ -38,10 +39,6 @@ export class Request {
 
   }
 
-  dateTimeLogger(): string {
-    return dateFormat(new Date(), "yyyy-mm-dd hh:mm:ss:l");
-  }
-
   hasSignature = (): boolean => {
     return (this.signature != undefined);
   }
@@ -75,7 +72,7 @@ export class Request {
 
           if (didDoc) {
             //cache-hit
-            console.log(this.dateTimeLogger() + ' got cache record for key ' + this.signature.creator);
+            console.log(dateTimeLogger() + ' got cache record for key ' + this.signature.creator);
             if (validateKyc) {
               if (!preVerifyDidSignature(didDoc, this, capability)) {
                 validator.addError("Signature failed pre verification " + this.signature.creator);
@@ -96,7 +93,7 @@ export class Request {
             resolve(validator);
           } else {
             //cache-miss
-            console.log(this.dateTimeLogger() + ' retrieve pubkey from blockchain');
+            console.log(dateTimeLogger() + ' retrieve pubkey from blockchain');
             axios.get(BLOCKCHAIN_URI_REST + 'did/getByDid/' + this.signature.creator)
               .then((response) => {
                 if (response.status == 200 && response.data.did != null) {
@@ -137,8 +134,8 @@ export class Request {
         })
         .catch((reason) => {
           // could not connect to cache, read from blockchain
-          console.log(this.dateTimeLogger() + ' cache unavailable ' + reason);
-          console.log(this.dateTimeLogger() + ' retrieve pubkey from blockchain');
+          console.log(dateTimeLogger() + ' cache unavailable ' + reason);
+          console.log(dateTimeLogger() + ' retrieve pubkey from blockchain');
           axios.get(BLOCKCHAIN_URI_REST + 'did/getByDid/' + this.signature.creator)
             .then((response) => {
               if (response.status == 200) {
