+++ src/handlers/RequestHandler.ts
@@ -8,7 +8,7 @@ import submitClaimProcessor from '../ixo/processor/SubmitClaimProcessor';
 import updateAgentStatusProcessor from '../ixo/processor/UpdateAgentStatusProcessor';
 
 import mq from '../MessageQ';
-import { dateTimeLogger } from '../ixo/common/shared';
+import { dateTimeLogger } from '../logger/Logger';
 
 export class RequestHandler {
