+++ src/ixo/common/shared.ts
@@ -1,5 +1,3 @@
-var dateFormat = require('dateformat');
-
 export const enum Status {
     created = "CREATED",
     pending = "PENDING",
@@ -8,7 +6,3 @@ export const enum Status {
     stopped = "STOPPED",
     payout = "PAIDOUT"
 }
-
-export function dateTimeLogger(): string {
-    return dateFormat(new Date(), "yyyy-mm-dd hh:mm:ss:l");
-  }
