+++ src/ixo/processor/ListClaimsProcessor.ts
@@ -1,7 +1,7 @@
 import { AbstractHandler } from '../../handlers/AbstractHandler';
 import { Claim } from '../model/ClaimModel';
 import { Request } from "../../handlers/Request";
-import { dateTimeLogger } from '../common/shared';
+import { dateTimeLogger } from '../../logger/Logger';
 
 export class ListClaimProcessor extends AbstractHandler {
