+++ src/ixo/processor/SubmitClaimProcessor.ts
@@ -2,7 +2,7 @@ import { AbstractHandler } from '../../handlers/AbstractHandler';
 import { Claim } from '../model/ClaimModel';
 import { ProjectStatus, IProjectStatusModel } from '../model/ProjectStatusModel';
 import { Request } from "../../handlers/Request";
-import { dateTimeLogger } from '../common/shared';
+import { dateTimeLogger } from '../../logger/Logger';
 
 export class SubmitClaimProcessor extends AbstractHandler {
