+++ src/ixo/processor/UpdateAgentStatusProcessor.ts
@@ -2,7 +2,7 @@ import { AbstractHandler } from '../../handlers/AbstractHandler';
 import { AgentStatus, IAgentStatusModel } from '../model/AgentStatusModel';
 import { ProjectStatus, IProjectStatusModel } from '../model/ProjectStatusModel';
 import { Request } from "../../handlers/Request";
-import { dateTimeLogger } from '../common/shared';
+import { dateTimeLogger } from '../../logger/Logger';
 
 export class UpdateAgentStatusProcessor extends AbstractHandler {
