+++ src/templates/TemplateUtils.ts
@@ -1,6 +1,7 @@
 import { GitUtils } from './GitUtils';
 import { Validator } from 'jsonschema';
 import Cache from '../Cache';
+import { dateTimeLogger } from '../logger/Logger';
 
 var dateFormat = require('dateformat');
 
@@ -12,10 +13,6 @@ export class TemplateUtils {
     this.gitUtils = new GitUtils();
   }
 
-  dateTimeLogger(): string {
-    return dateFormat(new Date(), "yyyy-mm-dd hh:mm:ss:l");
-  }
-
   getTemplateFromCache(templateType: string, name: string): Promise<string> {
     return new Promise((resolve: Function, reject: Function) => {
       var key = this.getCacheKey(templateType, name);
@@ -23,7 +20,7 @@ export class TemplateUtils {
         .then((template: string) => {
           if (template) {
             // cache-hit
-            console.log(this.dateTimeLogger() + ' got cache record for key ' + key);
+            console.log(dateTimeLogger() + ' got cache record for key ' + key);
             resolve(template);
           } else {
             resolve(this.getTemplateFromRegistry(templateType, name));
@@ -31,7 +28,7 @@ export class TemplateUtils {
         })
         .catch((reason) => {
           // cannot connect to cache; cache-miss
-          console.log(this.dateTimeLogger() + ' template registry failed ' + reason);
+          console.log(dateTimeLogger() + ' template registry failed ' + reason);
           resolve(this.getTemplateFromRegistry(templateType, name));
         });
     });
@@ -39,7 +36,7 @@ export class TemplateUtils {
 
   getTemplateFromRegistry(templateType: string, name: string): any {
     var template = this.constructTemplate(templateType, name);
-    console.log(this.dateTimeLogger() + ' load template contents from file');
+    console.log(dateTimeLogger() + ' load template contents from file');
     return this.gitUtils.loadFileContents(template)
       .then((templateContents: any) => {
         var res = JSON.parse(templateContents);
